requirements.txt
Flask==3.0.3
opencv-python==4.9.0.80
numpy==1.26.4
google-generativeai==0.7.1
moviepy==1.0.3